requirements.txt
openai
python-dotenv
pyyaml
nltk
langchain
